+++ pkg/grafana/grafana.go
@@ -20,12 +20,6 @@ const (
 
 var orgNotFoundError = errors.New("organization not found")
 
-var SharedOrg = Organization{
-	ID:        1,
-	Name:      "Shared Org",
-	TenantIDs: []string{"giantswarm"},
-}
-
 // We need to use a custom name for now until we can replace the existing datasources.
 var defaultDatasources = []Datasource{
 	{